refactor(scripts): Move dev scripts from runtime lib folder
