+++ scripts/api-changelog.js
@@ -28,7 +28,7 @@ if (fs.existsSync(apiSignatureFile)) {
 return Promise.resolve()
     .then(() => {
         return new Promise((resolve, reject) => {
-            const parsejs = path.resolve(parentDirectory, 'lib', 'codegen', 'parsejs.js');
+            const parsejs = path.resolve(parentDirectory, 'scripts', 'parsejs.js');
             const command = child_process.fork(parsejs, ['--format', 'APISignature', '--inputDir', path.resolve(parentDirectory, 'lib'), '--outputDir', parentDirectory]);
             command.on('exit', (code) => {
                 if (code !== 0) {
@@ -40,7 +40,7 @@ return Promise.resolve()
     })
     .then(() => {
         return new Promise((resolve, reject) => {
-            const changelog = path.resolve(parentDirectory, 'lib', 'tools', 'changelog.js');
+            const changelog = path.resolve(parentDirectory, 'scripts', 'tools', 'changelog.js');
             const command = child_process.fork(changelog, ['--api', apiSignatureFile, '--changelog', path.resolve(parentDirectory, 'changelog.txt')]);
             command.on('exit', (code) => {
                 if (code !== 0) {
